app/requirements.txt
fastapi==0.96.0
# numpy==1.23.1
loguru==0.7.0
torch==2.5.1
transformers==4.46.3
tqdm==4.66.5
# matplotlib==3.5.2
scikit-learn==1.0.2
uvicorn==0.29.0
python-dotenv==1.0.1
pinecone==5.4.0
google-cloud-storage==2.18.2
python-multipart==0.0.6
# torchvision==0.20.1
pillow==10.4.0